Adjust CloudSet to new matlab struct
Use "filter_trajectory" as default top level struct name. Adjust
structname default value in CloudSet and loadCloudTracks.
Adjust readMAT and storeMAT! to new name centrLatLon and the reversed
order of Lat in the first column and Lon in the second. Rename variable 
lonlat to latlon for consistency in the code.